dcp test fixed

+++ src/deps/iniparser/iniparser.h
@@ -195,7 +195,7 @@ int iniparser_getint(dictionary * d, const char * key, int notfound);
   slightly modified to return long instead of int.
  */
 /*--------------------------------------------------------------------------*/
-long iniparser_getlint(dictionary * d, const char * key, int notfound);
+int32_t iniparser_getlint(dictionary * d, const char * key, int notfound);
 
 /*-------------------------------------------------------------------------*/
 /**
